+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - New "if not" tag. A drop-in replacement for the standard `if` tag that supports logical `not` and grouping with parentheses.
 - Moved some module-level constants to static class variables for easier subclassing of tags.
-- Allow `ObjectChain` to contain `push`, `pop` and `size` properties.
+- Allow `ObjectChain` to contain `push`, `pop` and `size` properties. This is a **potentially breaking change** for early adopters writing custom tags using `context.scope.push` and `context.scope.pop`. The symbols `chainPush` and `chainPop` would be needed instead.
+- Enforce `maxContextDepth` when extended a render context scope as well as copying a `RenderContext`.
 
 ## Version 1.0.0-beta.1